requirements.txt
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
grpcio==1.34.0
grpcio-tools==1.34.0
idna==2.10
iniconfig==1.1.1
joblib==1.0.0
more-itertools==8.6.0
music21==6.3.0
packaging==20.8
pluggy==0.13.1
protobuf==3.14.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.2
webcolors==1.11.1
